requirements.txt
libsass==0.13.7
cssmin==0.2.0
jsmin==2.2.2
jsonschema==2.6.0
SQLAlchemy==1.2.2
Flask==0.12.2
Flask-Assets==0.12
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
Flask-Cors==3.0.3
Pillow==4.0.0
pyScss==1.3.5
PyYAML==3.12
WTForms==2.1
bcrypt==3.1.4
blinker==1.4
dotenv-python==0.0.1
elasticsearch==6.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
